Monkey-patch drush module for puppet 4
The puppet-drush module that the puppet-drupal module depends on has
some quirks on puppet 4. First, with the new ordering algorithm, it's
not guaranteed that the drush repo and the composer installation will
happen in the same order as before. We can fix that without touching the
drush module itself by creating resource orderings from the
drupal::drush class. Second, on puppet 3 the empty string '' was
interpreted as falsey, so you could expect `if '' {}` to evaluate as
false. On puppet 4, it's interpreted as truthy, and for the drush module
using '' as the default for the git_tag parameter causes the drush
$git_ref variable to be set to the empty string which causes the module
to stay on the master branch instead of checking out the 6.x branch.
There's no semantic difference between the git_branch and git_tag
parameter, so we'll just set git_tag to the same as git_branch.

Change-Id: I3acbda4edd546b4f145c9991f18ca80bb3e2171e